updateBalance desafio-6
