+++ semana14/projeto-labeBank/package.json
@@ -7,7 +7,8 @@
     "moment": "tsc && node ./build/moment.js",
     "start": "tsc && node ./build/index.js",
     "abrir-conta": "tsc && node ./build/accounts.js",
-    "saldo": "tsc && node ./build/operations.js",
+    "saldo": "tsc && node ./build/getBalance.js",
+    "update": "tsc && node ./build/updateBalance.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
