+++ semana14/projeto-labeBank/src/payBill.ts
@@ -5,52 +5,6 @@ import { Account, Transaction, TransactionsEnum } from "./types"
 
 var colors = require('colors/safe')
 
-const getBalance = (name: string, cpf: string): void => {
-    const accountsArray: Account[] = getAllAccounts()
-    
-    const accountIndex: number = accountsArray.findIndex(
-        (costumer => costumer.name === name && costumer.cpf === Number(cpf)) 
-    )
-   
-    if (accountIndex !== -1) {
-        const balance: number = accountsArray[accountIndex].balance
-        console.log(`Conta de ${accountsArray[accountIndex].name}\nSaldo: ${balance}`)
-        return
-    } else {
-        console.log(colors.red("Dados inválidos"))
-        return
-    }
-
-}
-
-const addBalance = (name: string, cpf: string, value: string): void => {
-    const accountsArray: Account[] = getAllAccounts()
-    
-    const accountIndex: number = accountsArray.findIndex(
-        (costumer => costumer.name === name && costumer.cpf === Number(cpf)) 
-    )
-    
-    if (accountIndex === -1) {
-        console.log(colors.red("Dados inválidos"))
-        return
-    } 
-
-    const transaction: Transaction = {
-        type: TransactionsEnum.ADD_BALANCE,
-        value: Number(value),
-        date: moment().unix(),
-        description: "Depósito em dinheiro",
-        completed: true,
-    }
-
-    accountsArray[accountIndex].accountStatement.push(transaction)
-    accountsArray[accountIndex].balance += transaction.value
-
-    writeToDatabase(accountsArray)
-
-    console.log(colors.green("Depósito realizado com sucesso"))
-}
-
 const payBill = (cpf: string, description: string, value: string, date?: string): void => {
     const accountsArray = getAllAccounts()
 
@@ -92,4 +46,5 @@ const payBill = (cpf: string, description: string, value: string, date?: string)
         console.log(colors.green("Pagamento realizado"))
     }    
 }
+
